Break up sklearn module into submodules
