+++ src/augury/sklearn/models.py
@@ -1,19 +1,15 @@
 """Classes and functions based on existing Scikit-learn functionality."""
 
 from typing import Sequence, Type, List, Union, Optional, Any, Tuple, Dict, Callable
-import re
 import copy
-from functools import partial, update_wrapper
 import warnings
 import tempfile
-import math
 
 import pandas as pd
 import numpy as np
 from sklearn.base import (
     BaseEstimator,
     RegressorMixin,
-    TransformerMixin,
     ClassifierMixin,
 )
 from sklearn.utils.metaestimators import _BaseComposition
@@ -22,82 +18,21 @@ from mypy_extensions import TypedDict
 from statsmodels.tsa.base.tsa_model import TimeSeriesModel
 from scipy.stats import norm
 from tensorflow import keras
-import tensorflow as tf
 
-from augury.types import R, T
+from augury.types import R
 from augury.nodes.base import _validate_required_columns
 from augury.nodes import common
+from augury.sklearn.metrics import bits_loss
 from augury.settings import TEAM_NAMES
 
 
-EloDictionary = TypedDict(
-    "EloDictionary",
-    {
-        "prev_team_elo_ratings": np.ndarray,
-        "current_team_elo_ratings": np.ndarray,
-        "year": int,
-        "round_number": int,
-    },
-)
-
-
-MATCH_INDEX_COLS = ["year", "round_number"]
-MATCH_COLS = ["date", "venue", "round_type"]
-OPPO_REGEX = re.compile("^oppo_")
-
-ELO_INDEX_COLS = ["home_team", "year", "round_number"]
-NULL_TEAM_NAME = "0"
-
-BASE_RATING = 1000
-K = 35.6
-ELO_X = 0.49
-M = 130
-HOME_GROUND_ADVANTAGE = 9
-S = 250
-SEASON_CARRYOVER = 0.575
-
-MATRIX_COLS = [
-    "year",
-    "round_number",
-    "home_team",
-    "home_prev_match_at_home",
-    "home_prev_match_oppo_team",
-    "home_prev_match_margin",
-    "away_team",
-    "away_prev_match_at_home",
-    "away_prev_match_oppo_team",
-    "away_prev_match_margin",
-]
-YEAR_IDX = MATRIX_COLS.index("year")
-ROUND_NUMBER_IDX = MATRIX_COLS.index("round_number")
-HOME_TEAM_IDX = MATRIX_COLS.index("home_team")
-AWAY_TEAM_IDX = MATRIX_COLS.index("away_team")
-PREV_AT_HOME_OFFSET = 1
-PREV_OPPO_OFFSET = 2
-PREV_MARGIN_OFFSET = 3
-
-YEAR_LVL = 1
-ROUND_NUMBER_LVL = 2
-
-# Got these default p, d, q values from a combination of statistical tests
-# (mostly for d and q) and trial-and-error (mostly for p)
-DEFAULT_ORDER = (6, 0, 1)
-# Minimum number of training years when using exogeneous variables
-# in a time-series model to avoid the error "On entry to DLASCL parameter number 4
-# had an illegal value". Arrived at through trial-and-error.
-MIN_YEARS_FOR_DLASCL = 3
-# Minimum number of training years when using exogeneous variables
-# in a time-series model to avoid the warning "HessianInversionWarning:
-# Inverting hessian failed, no bse or cov_params available".
-# Arrived at through trial-and-error.
-MIN_YEARS_FOR_HESSIAN = 6
-MIN_TIME_SERIES_YEARS = max(MIN_YEARS_FOR_DLASCL, MIN_YEARS_FOR_HESSIAN)
-# For regressors that might try to predict negative values or 0,
-# we need a slightly positive minimum to not get errors when calculating logarithms
-MIN_LOG_VAL = 1 * 10 ** -10
-LOSS = 0
-DRAW = 0.5
-WIN = 1
+# Default params for EloRegressor
+DEFAULT_K = 35.6
+DEFAULT_X = 0.49
+DEFAULT_M = 130
+DEFAULT_HOME_GROUND_ADVANTAGE = 9
+DEFAULT_S = 250
+DEFAULT_SEASON_CARRYOVER = 0.575
 
 
 class AveragingRegressor(_BaseComposition, RegressorMixin):
@@ -162,82 +97,55 @@ class AveragingRegressor(_BaseComposition, RegressorMixin):
             )
 
 
-class CorrelationSelector(BaseEstimator, TransformerMixin):
-    """Transformer for filtering out features that are less correlated with labels."""
-
-    def __init__(
-        self,
-        cols_to_keep: List[str] = [],
-        threshold: Optional[float] = None,
-        labels=pd.Series(),
-    ) -> None:
-        """Instantiate a CorrelationSelector transformer.
-
-        Params
-        ------
-        cols_to_keep: List of feature names to always keep in the data set.
-        threshold: Minimum correlation value (exclusive) for keeping a feature.
-        labels: Label values from the training data set for calculating
-            correlations.
-        """
-        self.threshold = threshold
-        self.labels = labels
-        self._cols_to_keep = cols_to_keep
-        self._above_threshold_columns = cols_to_keep
-
-    def transform(self, X: pd.DataFrame, _y=None) -> pd.DataFrame:
-        """Filter out features with weak correlation with the labels."""
-        return X[self._above_threshold_columns]
-
-    def fit(self, X: pd.DataFrame, y: Optional[pd.Series] = None) -> Type[T]:
-        """Calculate feature/label correlations and save high-correlation features."""
-        if not any(self.labels) and y is not None:
-            self.labels = y
-
-        assert any(
-            self.labels
-        ), "Need labels argument for calculating feature correlations."
-
-        data_frame = pd.concat([X, self.labels], axis=1).drop(self.cols_to_keep, axis=1)
-        label_correlations = data_frame.corr().fillna(0)[self.labels.name].abs()
-
-        if self.threshold is None:
-            correlated_columns = data_frame.columns
-        else:
-            correlated_columns = data_frame.columns[label_correlations > self.threshold]
-
-        self._above_threshold_columns = self.cols_to_keep + [
-            col for col in correlated_columns if col in X.columns
-        ]
-
-        return self
-
-    @property
-    def cols_to_keep(self) -> List[str]:
-        """List columns that never filtered out."""
-        return self._cols_to_keep
-
-    @cols_to_keep.setter
-    def cols_to_keep(self, cols_to_keep: List[str]) -> None:
-        """Set the list of columns to always keep.
-
-        Also resets the overall list of columns to keep to the given list.
-        """
-        self._cols_to_keep = cols_to_keep
-        self._above_threshold_columns = self._cols_to_keep
-
-
 class EloRegressor(BaseEstimator, RegressorMixin):
     """Elo regression model with a scikit-learn interface."""
 
+    EloDictionary = TypedDict(
+        "EloDictionary",
+        {
+            "previous_elo": np.array,
+            "current_elo": np.array,
+            "year": int,
+            "round_number": int,
+        },
+    )
+
+    ELO_INDEX_COLS = ["home_team", "year", "round_number"]
+    NULL_TEAM_NAME = "0"
+    BASE_RATING = 1000
+
+    # Constants for accessing data in the running_elo_ratings matrices
+    MATRIX_COLS = [
+        "year",
+        "round_number",
+        "home_team",
+        "home_prev_match_at_home",
+        "home_prev_match_oppo_team",
+        "home_prev_match_margin",
+        "away_team",
+        "away_prev_match_at_home",
+        "away_prev_match_oppo_team",
+        "away_prev_match_margin",
+    ]
+    YEAR_IDX = MATRIX_COLS.index("year")
+    ROUND_NUMBER_IDX = MATRIX_COLS.index("round_number")
+    HOME_TEAM_IDX = MATRIX_COLS.index("home_team")
+    AWAY_TEAM_IDX = MATRIX_COLS.index("away_team")
+    PREV_AT_HOME_OFFSET = 1
+    PREV_OPPO_OFFSET = 2
+    PREV_MARGIN_OFFSET = 3
+
+    YEAR_LVL = 1
+    ROUND_NUMBER_LVL = 2
+
     def __init__(
         self,
-        k=K,
-        x=ELO_X,
-        m=M,
-        home_ground_advantage=HOME_GROUND_ADVANTAGE,
-        s=S,
-        season_carryover=SEASON_CARRYOVER,
+        k=DEFAULT_K,
+        x=DEFAULT_X,
+        m=DEFAULT_M,
+        home_ground_advantage=DEFAULT_HOME_GROUND_ADVANTAGE,
+        s=DEFAULT_S,
+        season_carryover=DEFAULT_SEASON_CARRYOVER,
     ):
         """
         Instantiate an EloRegressor object.
@@ -259,13 +167,15 @@ class EloRegressor(BaseEstimator, RegressorMixin):
         self.home_ground_advantage = home_ground_advantage
         self.s = s
         self.season_carryover = season_carryover
-        self._running_elo_ratings = {
+        self._running_elo_ratings: self.EloDictionary = {
             "previous_elo": np.array([]),
             "current_elo": np.array([]),
             "year": 0,
             "round_number": 0,
         }
-        self._fitted_elo_ratings = copy.deepcopy(self._running_elo_ratings)
+        self._fitted_elo_ratings: self.EloDictionary = copy.deepcopy(
+            self._running_elo_ratings
+        )
         self._first_fitted_year = 0
 
         self._team_encoder = LabelEncoder()
@@ -273,17 +183,17 @@ class EloRegressor(BaseEstimator, RegressorMixin):
         # on the teams in the train set being a superset of those in the test set.
         # Have to add '0' to team names to account for filling in prev_match_oppo_team
         # for a new team's first match
-        self._team_encoder.fit(np.append(np.array(TEAM_NAMES), NULL_TEAM_NAME))
-        self._null_team = self._team_encoder.transform([NULL_TEAM_NAME])[0]
+        self._team_encoder.fit(np.append(np.array(TEAM_NAMES), self.NULL_TEAM_NAME))
+        self._null_team = self._team_encoder.transform([self.NULL_TEAM_NAME])[0]
 
     def fit(self, X: pd.DataFrame, _y: pd.Series = None) -> Type[R]:
         """Fit estimator to data."""
-        REQUIRED_COLS = set(ELO_INDEX_COLS) | set(MATRIX_COLS)
+        REQUIRED_COLS = set(self.ELO_INDEX_COLS) | set(self.MATRIX_COLS)
         _validate_required_columns(REQUIRED_COLS, X.columns)
 
         data_frame: pd.DataFrame = (
-            X.set_index(ELO_INDEX_COLS, drop=False)
-            .rename_axis([None] * len(ELO_INDEX_COLS))
+            X.set_index(self.ELO_INDEX_COLS, drop=False)
+            .rename_axis([None] * len(self.ELO_INDEX_COLS))
             .assign(
                 home_team=lambda df: self._team_encoder.transform(df["home_team"]),
                 away_team=lambda df: self._team_encoder.transform(df["away_team"]),
@@ -294,12 +204,12 @@ class EloRegressor(BaseEstimator, RegressorMixin):
                     df["away_prev_match_oppo_team"].astype(str)
                 ),
             )
-            .sort_index(level=[YEAR_LVL, ROUND_NUMBER_LVL], ascending=True)
+            .sort_index(level=[self.YEAR_LVL, self.ROUND_NUMBER_LVL], ascending=True)
         )
 
         self._reset_elo_state()
 
-        elo_matrix = (data_frame.loc[:, MATRIX_COLS]).to_numpy()
+        elo_matrix = (data_frame.loc[:, self.MATRIX_COLS]).to_numpy()
 
         for match_row in elo_matrix:
             self._update_current_elo_ratings(match_row)
@@ -315,7 +225,7 @@ class EloRegressor(BaseEstimator, RegressorMixin):
         Data set used for predictions must follow the training set chronologically.
         Otherwise, an error is raised to avoid making invalid predictions.
         """
-        REQUIRED_COLS = set(ELO_INDEX_COLS) | {"away_team"}
+        REQUIRED_COLS = set(self.ELO_INDEX_COLS) | {"away_team"}
 
         _validate_required_columns(REQUIRED_COLS, X.columns)
 
@@ -337,11 +247,11 @@ class EloRegressor(BaseEstimator, RegressorMixin):
                     df["away_prev_match_oppo_team"].astype(str)
                 ),
             )
-            .set_index(ELO_INDEX_COLS, drop=False)
-            .sort_index(level=[YEAR_LVL, ROUND_NUMBER_LVL], ascending=True)
+            .set_index(self.ELO_INDEX_COLS, drop=False)
+            .sort_index(level=[self.YEAR_LVL, self.ROUND_NUMBER_LVL], ascending=True)
         )
 
-        elo_matrix = (data_frame.loc[:, MATRIX_COLS]).to_numpy()
+        elo_matrix = (data_frame.loc[:, self.MATRIX_COLS]).to_numpy()
 
         if data_frame["year"].min() == self._first_fitted_year:
             self._reset_elo_state()
@@ -378,8 +288,8 @@ class EloRegressor(BaseEstimator, RegressorMixin):
         )
 
     def _calculate_current_elo_predictions(self, match_row: np.ndarray):
-        home_team = int(match_row[HOME_TEAM_IDX])
-        away_team = int(match_row[AWAY_TEAM_IDX])
+        home_team = int(match_row[self.HOME_TEAM_IDX])
+        away_team = int(match_row[self.AWAY_TEAM_IDX])
 
         self._update_current_elo_ratings(match_row)
 
@@ -398,20 +308,20 @@ class EloRegressor(BaseEstimator, RegressorMixin):
     # Assumes df sorted by year & round_number with ascending=True in order to calculate
     # correct Elo ratings
     def _update_current_elo_ratings(self, match_row: np.ndarray) -> None:
-        home_team = int(match_row[HOME_TEAM_IDX])
-        away_team = int(match_row[AWAY_TEAM_IDX])
+        home_team = int(match_row[self.HOME_TEAM_IDX])
+        away_team = int(match_row[self.AWAY_TEAM_IDX])
 
         self._update_prev_elo_ratings(match_row)
 
-        home_elo_rating = self._calculate_current_elo_rating(HOME_TEAM_IDX, match_row)
-        away_elo_rating = self._calculate_current_elo_rating(AWAY_TEAM_IDX, match_row)
+        home_elo_rating = self._calculate_current_elo_rating(self.HOME_TEAM_IDX, match_row)
+        away_elo_rating = self._calculate_current_elo_rating(self.AWAY_TEAM_IDX, match_row)
 
         self._running_elo_ratings["current_elo"][home_team] = home_elo_rating
         self._running_elo_ratings["current_elo"][away_team] = away_elo_rating
 
     def _update_prev_elo_ratings(self, match_row: np.ndarray):
-        match_year = match_row[YEAR_IDX]
-        match_round = match_row[ROUND_NUMBER_IDX]
+        match_year = match_row[self.YEAR_IDX]
+        match_round = match_row[self.ROUND_NUMBER_IDX]
 
         self._validate_consecutive_rounds(match_year, match_round)
 
@@ -430,15 +340,15 @@ class EloRegressor(BaseEstimator, RegressorMixin):
         if match_year != self._running_elo_ratings["year"]:
             self._running_elo_ratings["previous_elo"] = (
                 self._running_elo_ratings["previous_elo"] * self.season_carryover
-            ) + BASE_RATING * (1 - self.season_carryover)
+            ) + self.BASE_RATING * (1 - self.season_carryover)
 
             self._running_elo_ratings["year"] = match_year
 
     def _calculate_current_elo_rating(self, team_idx: int, match_row: np.ndarray):
         team = int(match_row[team_idx])
-        was_at_home = int(match_row[team_idx + PREV_AT_HOME_OFFSET])
-        prev_oppo_team = int(match_row[team_idx + PREV_OPPO_OFFSET])
-        prev_margin = match_row[team_idx + PREV_MARGIN_OFFSET]
+        was_at_home = int(match_row[team_idx + self.PREV_AT_HOME_OFFSET])
+        prev_oppo_team = int(match_row[team_idx + self.PREV_OPPO_OFFSET])
+        prev_margin = match_row[team_idx + self.PREV_MARGIN_OFFSET]
 
         # If a previous oppo team has the null team value, that means this is
         # the given team's first match, so they start with the default Elo rating
@@ -473,16 +383,16 @@ class EloRegressor(BaseEstimator, RegressorMixin):
     def _calculate_team_elo_rating(
         self, elo_rating: float, elo_prediction: float, margin: int
     ) -> float:
-        actual_outcome = self.x + 0.5 - self.x ** (1 + (margin / M))
+        actual_outcome = self.x + 0.5 - self.x ** (1 + (margin / self.m))
 
         return elo_rating + (self.k * (actual_outcome - elo_prediction))
 
     def _reset_elo_state(self):
         self._running_elo_ratings["previous_elo"] = np.full(
-            len(TEAM_NAMES) + 1, BASE_RATING
+            len(TEAM_NAMES) + 1, self.BASE_RATING
         )
         self._running_elo_ratings["current_elo"] = np.full(
-            len(TEAM_NAMES) + 1, BASE_RATING
+            len(TEAM_NAMES) + 1, self.BASE_RATING
         )
         self._running_elo_ratings["year"] = 0
         self._running_elo_ratings["round_number"] = 0
@@ -506,161 +416,22 @@ class EloRegressor(BaseEstimator, RegressorMixin):
         )
 
 
-class TeammatchToMatchConverter(BaseEstimator, TransformerMixin):
-    """Transformer for converting data frames to be organised by match."""
-
-    def __init__(self, match_cols=MATCH_COLS):
-        """
-        Instantiate a TeammatchToMatchConverter transformer.
-
-        Params
-        ------
-        match_cols (list of strings,
-            default=["date", "venue", "round_type"]):
-            List of match columns that are team neutral (e.g. round_number, venue).
-            These won't be renamed with 'home_' or 'away_' prefixes.
-        """
-        self.match_cols = match_cols
-        self._match_cols = list(set(match_cols + MATCH_INDEX_COLS))
-
-    def fit(self, _X, _y=None):
-        """Include for consistency with the Scikit-learn interface."""
-        return self
-
-    def transform(self, X: pd.DataFrame) -> pd.DataFrame:
-        """Transform data from being organised by team-match to match.
-
-        This means that input has two rows per match (one row each
-        for home and away teams), and output has one row per match (with separate
-        columns for home and away team data).
-        """
-        self._validate_required_columns(X)
-
-        return (
-            pd.concat(
-                [
-                    self._match_data_frame(X, at_home=True),
-                    self._match_data_frame(X, at_home=False),
-                ],
-                axis=1,
-            )
-            .reset_index()
-            .set_index(["home_team"] + MATCH_INDEX_COLS, drop=False)
-            .rename_axis([None] * (len(MATCH_INDEX_COLS) + 1))
-            .sort_index()
-        )
-
-    def _validate_required_columns(self, data_frame: pd.DataFrame):
-        required_cols: List[str] = ["team", "oppo_team", "at_home"] + self._match_cols
-
-        _validate_required_columns(required_cols, data_frame.columns)
-
-    def _match_data_frame(
-        self, data_frame: pd.DataFrame, at_home: bool = True
-    ) -> pd.DataFrame:
-        home_index = "team" if at_home else "oppo_team"
-        away_index = "oppo_team" if at_home else "team"
-        # We drop oppo stats cols, because we end up with both teams' stats per match
-        # when we join home and away teams. We keep 'oppo_team' and add the renamed column
-        # to the index for convenience
-        oppo_stats_cols = [
-            col
-            for col in data_frame.columns
-            if re.match(OPPO_REGEX, col) and col != "oppo_team"
-        ]
-
-        return (
-            data_frame.query(f"at_home == {int(at_home)}")
-            # We index match rows by home_team, year, round_number
-            .rename(columns={home_index: "home_team", away_index: "away_team"})
-            .drop(["at_home"] + oppo_stats_cols, axis=1)
-            # We add all match cols to the index, because they don't affect the upcoming
-            # concat, and it's easier than creating a third data frame for match cols
-            .set_index(["home_team", "away_team"] + self._match_cols)
-            .rename(columns=self._replace_col_names(at_home))
-            .sort_index()
-        )
-
-    @staticmethod
-    def _replace_col_names(at_home: bool):
-        team_label = "home" if at_home else "away"
-        oppo_label = "away" if at_home else "home"
-
-        return (
-            lambda col: col.replace("oppo_", f"{oppo_label}_", 1)
-            if re.match(OPPO_REGEX, col)
-            else f"{team_label}_{col}"
-        )
-
-
-class ColumnDropper(BaseEstimator, TransformerMixin):
-    """Transformer that drops named columns from data frames."""
-
-    def __init__(self, cols_to_drop: List[str] = []):
-        """Instantiate a ColumbnDropper transformer.
-
-        Params
-        ------
-        cols_to_drop: List of column names to drop.
-        """
-        self.cols_to_drop = cols_to_drop
-
-    def fit(self, _X, y=None):  # pylint: disable=unused-argument
-        """Include for consistency with Scikit-learn interface."""
-        return self
-
-    def transform(self, X: pd.DataFrame) -> pd.DataFrame:
-        """Drop the given columns from the data."""
-        return X.drop(self.cols_to_drop, axis=1, errors="ignore")
-
-
-class DataFrameConverter(BaseEstimator, TransformerMixin):
-    """Transformer that converts numpy arrays into DataFrames with named axes.
-
-    Resulting data frame is assigned named columns and indices per the initial data sets
-    passed to fit/predict. This is mostly for cases when classes from packages
-    convert DataFrames to numpy arrays without asking, and later transformers depend
-    on named indices/columns to work.
-    """
-
-    def __init__(
-        self,
-        columns: Optional[Union[List[str], pd.Index]] = None,
-        index: Optional[Union[List[str], pd.Index]] = None,
-    ):
-        """Instantiate a DataFrameConverter transformer.
-
-        Params
-        ------
-        columns: List of column names or a pd.Index to assign as columns.
-        index: List of row names or a pd.Index to assign as the index.
-        """
-        self.columns = columns
-        self.index = index
-
-    def fit(self, X, y=None):  # pylint: disable=unused-argument
-        """Include for consistency with Scikit-learn interface."""
-        return self
-
-    def transform(self, X: Union[pd.DataFrame, np.array]):
-        """Convert data into a pandas DataFrame with the given columns and index."""
-        if self.columns is not None:
-            assert X.shape[1] == len(self.columns), (
-                f"X must have the same number of columns {X.shape[1]} "
-                f"as self.columns {len(self.columns)}."
-            )
-
-        if self.index is not None:
-            assert X.shape[0] == len(self.index), (
-                f"X must have the same number of rows {X.shape[0]} "
-                f"as indicated by self.index {len(self.index)}."
-            )
-
-        return pd.DataFrame(X, columns=self.columns, index=self.index)
-
-
 class TimeSeriesRegressor(BaseEstimator, RegressorMixin):
     """Wrapper class with Scikit-learn API for regressors from statsmodels.tsa."""
+
+    # Got these default p, d, q values from a combination of statistical tests
+    # (mostly for d and q) and trial-and-error (mostly for p)
+    DEFAULT_ORDER = (6, 0, 1)
+    # Minimum number of training years when using exogeneous variables
+    # in a time-series model to avoid the error "On entry to DLASCL parameter number 4
+    # had an illegal value". Arrived at through trial-and-error.
+    MIN_YEARS_FOR_DLASCL = 3
+    # Minimum number of training years when using exogeneous variables
+    # in a time-series model to avoid the warning "HessianInversionWarning:
+    # Inverting hessian failed, no bse or cov_params available".
+    # Arrived at through trial-and-error.
+    MIN_YEARS_FOR_HESSIAN = 6
+    MIN_TIME_SERIES_YEARS = max(MIN_YEARS_FOR_DLASCL, MIN_YEARS_FOR_HESSIAN)
 
     def __init__(
         self,
@@ -722,7 +493,7 @@ class TimeSeriesRegressor(BaseEstimator, RegressorMixin):
         # arbitrary, but were arrived at after a fair bit of trial-and-error.
         p_param = (
             self.order[0]
-            if n_train_years >= MIN_TIME_SERIES_YEARS
+            if n_train_years >= self.MIN_TIME_SERIES_YEARS
             else min(4, self.order[0])
         )
         order_param = (p_param, *self.order[1:])
@@ -775,65 +546,6 @@ class TimeSeriesRegressor(BaseEstimator, RegressorMixin):
 
     def _exog_arg(self, data_frame: pd.DataFrame) -> Optional[np.array]:
         return data_frame[self.exog_cols].to_numpy() if any(self.exog_cols) else None
-
-
-def _positive_pred_tensor(y_pred):
-    return tf.where(
-        tf.math.less_equal(y_pred, tf.constant(0.0)), tf.constant(MIN_LOG_VAL), y_pred
-    )
-
-
-def _log2(x):
-    return tf.math.divide(
-        tf.math.log(_positive_pred_tensor(x)), tf.math.log(tf.constant(2.0))
-    )
-
-
-def _draw_bits_tensor(y_pred):
-    return tf.math.add(
-        tf.constant(1.0),
-        tf.math.scalar_mul(
-            tf.constant(0.5),
-            _log2(tf.math.multiply(y_pred, tf.math.subtract(tf.constant(1.0), y_pred))),
-        ),
-    )
-
-
-def _win_bits_tensor(y_pred):
-    return tf.math.add(tf.constant(1.0), _log2(y_pred))
-
-
-def _loss_bits_tensor(y_pred):
-    return tf.math.add(
-        tf.constant(1.0), _log2(tf.math.subtract(tf.constant(1.0), y_pred))
-    )
-
-
-# Raw bits calculations per http://probabilistic-footy.monash.edu/~footy/about.shtml
-def bits_loss(y_true, y_pred):
-    """Loss function for Tensorflow models based on the bits metric."""
-    y_true_f = tf.cast(y_true, tf.float32)
-    y_pred_win = y_pred[:, -1:]
-
-    # We adjust bits calculation to make a valid ML error formula such that 0
-    # represents a correct prediction, and the further off the prediction
-    # the higher the error value.
-    return K_backend.mean(
-        tf.where(
-            tf.math.equal(y_true_f, tf.constant(0.5)),
-            tf.math.scalar_mul(tf.constant(-1.0), _draw_bits_tensor(y_pred_win)),
-            tf.where(
-                tf.math.equal(y_true_f, tf.constant(1.0)),
-                tf.math.subtract(tf.constant(1.0), _win_bits_tensor(y_pred_win)),
-                tf.math.add(
-                    tf.constant(1.0),
-                    tf.math.scalar_mul(
-                        tf.constant(-1.0), _loss_bits_tensor(y_pred_win)
-                    ),
-                ),
-            ),
-        ),
-    )
 
 
 class KerasClassifier(BaseEstimator, ClassifierMixin):
@@ -971,246 +683,3 @@ class KerasClassifier(BaseEstimator, ClassifierMixin):
         d = {value: key for value, key in state.items() if key != "model_str"}
         d.update({"model": model})
         self.__dict__ = d # pylint: disable=attribute-defined-outside-init
-
-
-def _calculate_team_margin(team_margin, oppo_margin):
-    # We want True to be 1 and False to be -1
-    team_margin_multiplier = ((team_margin > oppo_margin).astype(int) * 2) - 1
-
-    return (
-        pd.Series(
-            ((team_margin.abs() + oppo_margin.abs()) / 2) * team_margin_multiplier
-        )
-        .reindex(team_margin.index)
-        .sort_index()
-    )
-
-
-def _calculate_match_accuracy(X, y_true, y_pred):
-    """Scikit-learn metric function for calculating tipping accuracy."""
-    team_match_data_frame = X.assign(y_true=y_true, y_pred=y_pred)
-    home_match_data_frame = team_match_data_frame.query("at_home == 1").sort_index()
-    away_match_data_frame = (
-        team_match_data_frame.query("at_home == 0")
-        .set_index(["oppo_team", "year", "round_number"])
-        .rename_axis([None, None, None])
-        .sort_index()
-    )
-
-    home_margin = _calculate_team_margin(
-        home_match_data_frame["y_true"], away_match_data_frame["y_true"]
-    )
-    home_pred_margin = _calculate_team_margin(
-        home_match_data_frame["y_pred"], away_match_data_frame["y_pred"]
-    )
-
-    return (
-        # Any zero margin (i.e. a draw) is counted as correct per usual tipping rules.
-        # Predicted margins should never be zero, but since we don't want to encourage
-        # any wayward models, we'll count a predicted margin of zero as incorrect
-        ((home_margin >= 0) & (home_pred_margin > 0))
-        | ((home_margin <= 0) & (home_pred_margin < 0))
-    ).mean()
-
-
-def create_match_accuracy(X):
-    """Return Scikit-learn metric function for calculating tipping accuracy."""
-    return update_wrapper(
-        partial(_calculate_match_accuracy, X), _calculate_match_accuracy
-    )
-
-
-def match_accuracy_scorer(estimator, X, y):
-    """Scikit-learn scorer function for calculating tipping accuracy of an estimator."""
-    y_pred = estimator.predict(X)
-
-    return _calculate_match_accuracy(X, y, y_pred)
-
-
-def _positive_pred(y_pred):
-    return np.maximum(y_pred, np.repeat(MIN_LOG_VAL, len(y_pred)))
-
-
-def _draw_bits(y_pred):
-    return 1 + (0.5 * np.log2(_positive_pred(y_pred * (1 - y_pred))))
-
-
-def _win_bits(y_pred):
-    return 1 + np.log2(_positive_pred(y_pred))
-
-
-def _loss_bits(y_pred):
-    return 1 + np.log2(_positive_pred(1 - y_pred))
-
-
-# Raw bits calculations per http://probabilistic-footy.monash.edu/~footy/about.shtml
-def _calculate_bits(y_true, y_pred):
-    return np.where(
-        y_true == DRAW,
-        _draw_bits(y_pred),
-        np.where(y_true == WIN, _win_bits(y_pred), _loss_bits(y_pred)),
-    )
-
-
-def bits_scorer(
-    estimator: BaseEstimator,
-    X: Union[pd.DataFrame, np.ndarray],
-    y: pd.Series,
-    proba=False,
-    n_years=1,
-) -> float:
-    """Scikit-learn scorer for the bits metric.
-
-    Mostly for use in calls to cross_validate. Calculates a score
-    based on the the model's predicted probability of a given result. For this metric,
-    higher scores are better.
-
-    We simplify calculations by using Numpy math functions. This has the benefit
-    of not require a lot of reshaping based on categorical features, but gives
-    final values that deviate a little from what is correct, because this scorer
-    calculates bits per team-match combination rather than per match,
-    which is how the official bits score will be calculated.
-
-    Params
-    ------
-    estimator: The estimator being scored.
-    X: Model features.
-    y: Model labels.
-    proba: Whether to use the `predict_proba` method to get predictions.
-    """
-
-    try:
-        y_pred = estimator.predict_proba(X)[:, -1] if proba else estimator.predict(X)
-    # TF/Keras models don't use predict_proba, so for classifiers, we pass proba=True,
-    # then rescue and call predict.
-    except AttributeError:
-        if proba:
-            y_pred = estimator.predict(X)[:, -1]
-        else:
-            raise
-
-    if isinstance(X, pd.DataFrame) and "year" in X.columns:
-        n_years = X["year"].drop_duplicates().count()
-
-    # For tipping competitions, bits are summed across the season.
-    # We divide by number of seasons for easier comparison with other models.
-    # We divide by two to get a rough per-match bits value.
-    return _calculate_bits(y, y_pred).sum() / n_years / 2
-
-
-def _draw_bits_hessian(y_pred):
-    return (y_pred ** 2 - y_pred + 0.5) / (
-        math.log(2) * y_pred ** 2 * (y_pred - 1) ** 2
-    )
-
-
-def _win_bits_hessian(y_pred):
-    return 1 / (math.log(2) * y_pred ** 2)
-
-
-def _loss_bits_hessian(y_pred):
-    return 1 / (math.log(2) * (1 - y_pred) ** 2)
-
-
-def _bits_hessian(y_true, y_pred):
-    return np.where(
-        y_true == DRAW,
-        _draw_bits_hessian(y_pred),
-        np.where(y_true == WIN, _win_bits_hessian(y_pred), _loss_bits_hessian(y_pred),),
-    )
-
-
-def _draw_bits_gradient(y_pred):
-    return (y_pred - 0.5) / (math.log(2) * (y_pred - y_pred ** 2))
-
-
-def _win_bits_gradient(y_pred):
-    return -1 / (math.log(2) * y_pred)
-
-
-def _loss_bits_gradient(y_pred):
-    return 1 / (math.log(2) * (1 - y_pred))
-
-
-def _bits_gradient(y_true, y_pred):
-    return np.where(
-        y_true == DRAW,
-        _draw_bits_gradient(y_pred),
-        np.where(
-            y_true == WIN, _win_bits_gradient(y_pred), _loss_bits_gradient(y_pred),
-        ),
-    )
-
-
-def bits_objective(y_true, y_pred) -> Tuple[np.array, np.array]:
-    """Objective function for XGBoost estimators.
-
-    The gradient and hessian formulas are based on the formula for the bits error
-    function rather than the bits metric to make the math more consistent
-    with other objective and error functions.
-
-    Params
-    ------
-    y_true [array-like, (n_observations,)]: Data labels.
-    y_pred [array-like, (n_observations, n_label_classes)]: Model predictions.
-        In the case of binary classification, the shape is (n_observations,)
-
-    Returns
-    -------
-    gradient, hessian [tuple of array-like, (n_observations * n_classes,)]:
-        gradient function is the derivative of the loss function, and hessian function
-        is the derivative of the gradient function.
-    """
-    # Since y_pred can be 1- or 2-dimensional, we should only reshape y_true
-    # when the latter is the case.
-    y_true_matrix = (
-        y_true.reshape(-1, 1) if len(y_true.shape) != len(y_pred.shape) else y_true
-    )
-
-    return (
-        _bits_gradient(y_true_matrix, y_pred).flatten(),
-        _bits_hessian(y_true_matrix, y_pred).flatten(),
-    )
-
-
-def _bits_error(y_true, y_pred):
-    # We adjust bits calculation to make a valid ML error formula such that 0
-    # represents a correct prediction, and the further off the prediction
-    # the higher the error value.
-    return np.where(
-        y_true == DRAW,
-        -1 * _draw_bits(y_pred),
-        np.where(y_true == WIN, 1 - _win_bits(y_pred), 1 + (-1 * _loss_bits(y_pred)),),
-    )
-
-
-def bits_metric(y_pred, y_true_matrix) -> Tuple[str, float]:
-    """Metric function for internal model evaluation in XGBoost estimators.
-
-    Note that the order of params, per the xgboost documentation, is y_pred, y_true
-    as opposed to the usual y_true, y_pred for Scikit-learn metric functions.
-
-    Params
-    ------
-    y_pred: Model predictions.
-    y_true: Data labels.
-
-    Returns
-    -------
-    Tuple of the metric name and mean bits error.
-    """
-    y_true = y_true_matrix.get_label()
-
-    return "mean_bits_error", _bits_error(y_true, y_pred).mean()
-
-
-def year_cv_split(X, year_range):
-    """Split data by year for cross-validation for time-series data.
-
-    Makes data from each year in the year_range a test set per split, with data
-    from all earlier years being in the train split.
-    """
-    return [
-        ((X["year"] < year).to_numpy(), (X["year"] == year).to_numpy())
-        for year in range(*year_range)
-    ]
